Resource beliefs are pretty golden now!

+++ rover/rover/src/main/java/rover/shared/reasoning/APercept.java
@@ -1,11 +1,9 @@
 package rover.shared.reasoning;
 
-import rover.PollResult;
 import rover.ScanItem;
 import rover.shared.practical.RoverOffset;
 
 import java.util.ArrayList;
-import java.util.Collection;
 
 /**
  * Created by rachelcabot on 17/10/2016.
@@ -16,9 +14,9 @@ public abstract class APercept {
 
     public abstract RoverOffset getMyPosition();
 
-    public abstract int getWorldWidth();
+    public abstract double getWorldWidth();
 
-    public abstract int getWorldHeight();
+    public abstract double getWorldHeight();
 
     public abstract int getCurrentLoad();
